Refactor todo list component and add NgRx store
The 'TodoListComponent', along with its HTML and CSS, has been refactored and moved to the feature directory. In addition, the 'NgRx' store and effects have been included in the 'package.json' and 'package-lock.json' files. The system directory has been replaced with a more functional structure within the features directory. Tests for the component have been relocated accordingly.